attention_viz/requirements.txt
# Medical VLM Analysis Dependencies for Google Colab
# Optimized for Colab environment with modern package versions

# Core ML frameworks
torch>=2.4.0
torchvision>=0.19.0

# Transformers with Gemma3 support
transformers>=4.56.1

# Acceleration and quantization
accelerate>=0.34.2
bitsandbytes>=0.43.2

# Image processing
opencv-python>=4.10.0
pillow>=10.4.0

# Scientific computing
numpy>=1.24.0
scipy>=1.13.0
matplotlib>=3.9.0

# Additional utilities
einops>=0.8.0
gradio>=4.44.0